reorganized upload files
